requirements.txt
streamlit==1.31.1
yfinance==0.2.36
plotly==5.19.0
requests==2.31.0
python-dotenv==1.0.1
